pyproject.toml
[tool.poetry]
name = "test-app"
version = "0.1.0"
description = ""
authors = ["Мария Пирогова <mayefimenko@gmail.com>"]
readme = "README.md"
packages = [{include = "test_app"}]

[tool.poetry.dependencies]
python = "3.10.11"
django = "3.2"
djangorestframework = "^3.14.0"
black = "23.3.0"
psycopg2 = "^2.9.6"
drf-yasg = "^1.21.6"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 79